Split out viewer state and tools into separate files.
